changed it to have two exported functions as api.ts instead, this is probably the best it's gonna be to avoid looping dependencies
